refactor(game): split 745-line setup page into focused files under 250 lines
Reduces app/game/setup/page.tsx from 745 lines to 97 by extracting:

- 6 presentation components into app/game/setup/_components/
  - ExplorePanel — exploration phase UI (batch input + reveal log)
  - DistributePanel — distribute phase UI (counters + caste cards + tile grid)
  - CastePickCard — caste card with units/spells/buildings (with
    inline UnitLine/SpellLine/Stat helpers + topKey)
  - RevealLogList — field-report log
  - SetupGates — Loading/SignIn/Enlist/Complete screens

- 4 lib modules into app/game/setup/_lib/
  - use-setup-data — player + tiles state + refresh
  - use-setup-actions — callApi (single action) + runExploreBatch
  - types — PlayerResponse, RevealLog, BatchProgress
  - constants — CASTES, DISTRIBUTABLE, CASTE_PRESENTATION, RARITY_COLORS

The _components and _lib prefixes are Next.js private folders, so the new
structure does not introduce any new routes. page.tsx is now a thin gate
that picks one of {Loading, SignIn, Enlist, ExplorePanel, DistributePanel,
SetupCompleteScreen} based on the player's phase.

No behavior change. All 1581 tests pass; tsc clean; lint clean.
