+++ app/game/setup/_lib/use-setup-data.ts
@@ -1,0 +1,67 @@
+/**
+ * Copyright (C) 2026 Cursor Boston
+ * This file is part of Cursor Boston, licensed under GPL-3.0.
+ * See LICENSE file for details.
+ */
+
+"use client";
+
+import { useCallback, useEffect, useState } from "react";
+import { useAuth } from "@/contexts/AuthContext";
+import type { GamePlayer, MapTile } from "@/lib/game/types";
+import type { PlayerResponse } from "./types";
+
+/**
+ * Player + tiles state for the setup page. The setup flow only needs
+ * the casting player's record — no world cache, no border tiles. Every
+ * action handler refetches via `refresh()` because all setup actions
+ * affect player state (turns spent, tile types, caste).
+ */
+export function useSetupData() {
+  const { user, loading: authLoading } = useAuth();
+  const [player, setPlayer] = useState<GamePlayer | null>(null);
+  const [tiles, setTiles] = useState<MapTile[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  const refresh = useCallback(async () => {
+    setError(null);
+    if (!user) {
+      setLoading(false);
+      return;
+    }
+    try {
+      const token = await user.getIdToken();
+      const res = await fetch("/api/game/player", {
+        headers: { Authorization: `Bearer ${token}` },
+      });
+      const data = (await res.json()) as PlayerResponse;
+      if (!data.success) throw new Error(data.error ?? "Failed to load player");
+      setPlayer(data.player);
+      setTiles(data.tiles ?? []);
+    } catch (e) {
+      setError(e instanceof Error ? e.message : "Failed to load player");
+    } finally {
+      setLoading(false);
+    }
+  }, [user]);
+
+  useEffect(() => {
+    if (authLoading) return;
+    // eslint-disable-next-line react-hooks/set-state-in-effect -- fetch-on-mount, state set inside async callback
+    refresh();
+  }, [authLoading, refresh]);
+
+  return {
+    user,
+    authLoading,
+    player,
+    tiles,
+    loading,
+    error,
+    setError,
+    refresh,
+  };
+}
+
+export type SetupData = ReturnType<typeof useSetupData>;
